House Construction and Door Demolition
Made a simple script to create a house. Just makes the outer wall, a
floor, and a door right now. Should be able to extend this.

The axe now chops down doors it comes into contact with.
Leaning towards tools being triggers and having no gravity. Seems to make
it easier to work with. I don't know how I will place them in the world
when the players aren't using them yet...