Tools Force Movement
Added some support for forcing the player to move when using tools
The Player class subscribes to actions in the tool class.
When the tool wants to force the playerr to move, it calls ForcedMovement.
The player then swaps its movement style to match that tool. The player
has all the code for the movement styles. Currently, it is just forced
movement from the axe. It calls methods in the Tool class to get what
speed it should move and turn at. The tool can called a
ToolFinishedAction when it is finished forcing movement. The player
listens for this then swaps back to normal movement.

This seems to work well enough. I would like the movement logic to be in
the Tool class, but I'm not sure how to do that yet.
I would like more complex movement like a hose swaying back and forth in a
sort of arc pulling the character all over the place.
Creating a new Tool with movement currently requires:
Creating a new script, extending Tool.
Adding a new field to the ToolType enum.
Writing the code for the movement in the player class.
Tweaking numbers in the new tool subclass in the editor.

Potential solution:
Player class calls a new function Tool.MoveTowards instead of
Vector3.MoveTowards. That way each tool could tailor the different
movement.

House
I found out I could set the parent of an instantiated object when I
instantiate it(instead of on a new line) so I did that.