Extinguisher Fuel
Player can only use the extinguisher for so long before it runs out of
fuel. The extinguisher also pushes the player only a little off the start,
then ramps up the force. Once the extinguisher is out of fuel, it doens't
do anything.

Implemented the axe's custom MoveTowards function. It accesses its
parent's forward vector (the player's hand). I don't know if this is
better than the player just passing it into the function in the first
place or not.