Burning House
The house now starts on fire and the fire spreads to nearby blocks.

Each piece of the house is now a BuildingBlock class. This class handles
the burning and spreading of the fire. Different kinds of blocks can take
longer to burn than others. Nearby fire burns a BuildingBlock. After it
has burned for its fuelSeconds, it lights on fire. After that many
fuelSeconds again, the fire turns into a big fire. The big fire can then
burn its neighbours. After fuelSeconds a third time, the block is
considered destroyed by fire. Currently it stops spreading fire at this
point. This may change.

The house stores all blocks under a gameobject it creates so t caan delete
and remake the house easily.