Put Out Fire
Added a method to BuildingBlock to put out the fire. Nothing calls it yet
so it's not tested.

House
Added a method to rotate walls and doors. Will iterate through the grid
and find walls and their neighbours and rotate walls to be all in a line.
This isn't a priority until I have some wall models that are thin and not
cubes. But then the walls will need a floor beneath them so that is a
different problem.