Spray Cone to Extinguisher
Extinguisher now fires out rays in a cone in front of it.
The spray distance and cone arc can be adjusted in the editor.
The rays stop when they hit a wall, but right now the fire is inside the
walls... I don't want the rays to just ignore walls. Maybe the fire
collider just needs to be larger than the wall.