Particles on Extinguisher
The fire extinguisher emits particles when it is spraying. They are only
visual, they don't do anything. Also they are purple because they have no
texture.

Axe turns its trigger on and off when swinging down and swinging up. Makes
it more likely to break things in front of it. Before, the trigger would
never leave the object to be able to enter it again.