Added Victim Tool
The victim is a tool because the player can pick them up and press the use
button to toss them

Needed to make an accessor method in the player that the tool can call to
tell the player to force dropping it.
The victim tells the player to drop it when the player uses the victim
tool to throw the victim.
Victim has a solid collider and uses gravity to make the toss method easy.
Because the player also has gravity, this solves the old issue of having
solid colliders. When the player picks up the tool, the solid collider may
push the player. Without gravity, this pushes the player forever. With
gravity, that push goes away due to friction with the ground.

Moved where the player sets the ToolFinished and ToolStarted actions. Now
they are set right when the player picks up the object instead of when the
player uses them. The tools still have control over when they are fired so
that doesn't affect the other tools. It only lets the victim tool call
ForcedMovement as soon as it is picked up (it slows the player movement
because they are heavy).