Changed Tool Structure
When the player picks up a tool, they only pick up the visual part of the
tool. They no longer pick up a tool with a rigidbody. The rigidbody
component is left behind and hidden. When the player drops the tool, it is
reparented to the rigidbody and the rigidbody is unhidden and moved to the
correct location.

Having 2 rigidbodies conected to the same game object (even as a child)
causes problems. When the player would run into a wall, the tool would
keep moving forward.

The rigidbody component, a solid box collider and trigger sphere collider
make up the parent of the tool. Rigibody for gravity, solid box to not go
through the floor, and sphere for collision to let the player pick it up.
The child object has the visual model and the tool script. Also a trigger
box for the axe so it can break blocks.

Changed some GetComponents to now be GetComponentInParent or
GetComponentInChild to access things in their new spot.