Tool Drop Bugs
Trying to drop the axe while using it had unintended side effects.
Fixed so the player checks if they can drop the object before they do. For
the axe, they can't drop it during the swinging. Trying to drop the axe
during swinging will just fail to do anything

If dropping the extinguisher while spraying, the player would keep the
spraying movement after they dropped the extinguisher.
Fix: Extinguisher now calls ToolFinished() before it calls Drop()
Now the player will drop the extinguisher and stop moving.

Possible bug:
After using the victim to throw it, if the player immediately tries to
pick it up again, what happens? Does the player go flying with the victim?
I tried this a couple times and couldn't seem to do it. It might be
possible, might not. The collider may already be above the player when the
victim is thrown. Things may change if the victim is on a different spot
relative to the player or the pickup raycast code gets changed.