Fire Particles
When floor, walls, and doors are on fire, they emit fire particles.
I need to tone it down a bit when it's a small fire and maybe even tone it
down a bit for the big fire.
The particles also stay on after the tile has been destroyed by the fire.
It should change to embers or smoke at that point.