Camera Follow
Made a camera controler script. This will handle camera movement. The game
manager can give it places to focus on.
Currently, the game manager tells it to focus on the average position of
the players.

Also made a tile texture for the floor tiles. This will be swapped out for
something better. It was just to see if the tile helped tell where the
fire is, but it doesn't really seem to.