Victims to be Rescued
House now spawns with victims randomly strewn about. Victims will only
spawn on floor tiles.

All vicitms need to make it past a safe threshold to be considered safe.
Currently, this is just bringing victims down past z == 0.

You might be able to just toss 1 victim across the line 3 times to get it to count for all of them. But maybe not. The coroutine that checks their position
should be off once it crosses that line.

The checking if safe may be a little finicky. When the tools are picked up
and dropped, the rigidbody object (their parent) is turned on and off. If
the Victim script object is turned off, the coroutine checking its
position is also stopped. It works if the rigidbocy object is active when
the victim script is attached to it.