Collision Adjustments for Fire and Tools
Made the colliders for fire larger in the walls and door. This should make
it easier to hit the fire when it's inside of walls and doors. Before,
just a small chunk of the fire's capsule stuck out each face of the wall.
Now almost all of the fire is outside the wall cube.
The walls need colliders so players don't walk through them.
The extinguisher can't ignore walls because I don't want to shoot through
walls.

Tools are now on a tool layer. This should make pickup easier. Before,
the fire capsule could block you from being able to pick up tool that were
on the other side.
Also raised the collider on the extinguisher. The player could only hit
the top of its sphere. Now it should be at the right height to hit the
middle of the sphere.