Players On Fire
The players can be set on fire. Currently the only way to set them on fire
is by pressing fire3. In the future, being near fire in the house will set
them on fire.

When players are set on fire, they drop the tool they are using. When the
tool calls ToolFinished, player.ToolFinished() runs. In this method, the
player checks if they are on fire. If they are, they call pickup() to drop
the tool. This is for the axe mostly. The axe can't be dropped while in
use so the player just checks if it is on fire when the axe finished its
use and then drops it. Some calls needed to be reordered so as to not
cause an infinite loop of calls.

Extinguisher:
Checks if what it is hitting either has a BuildingBlock component in its
parent or a Player component in its parent to put out fire.

GameplayManager:
Keyboard input steals player4 instead of player 1. Otherwise the first
controller would do nothing if using the keyboard. You would still need to
plug in 2 controllers to get the second player to move though.

Player:
The player can start on fire.
When the player is on fire, they drop all tools and cannot pick up new
tools. They can press use to lay down on the ground to stop, drop, and
roll. They can only move left to right once they do this. If they move far
enough from their start position, it's considered a roll. x number of
rolls to put out the fire. The player can then press use again to get up
and resume all activity normally. Pressing a button to stand up again is a
little awkward and may need to be changed.
The fire extinguisher can also be used to put out the fire. It uses a
similar method to the buildingblocks, but the fire on the player doesn't
grow over time. This is proabably a bad practise. Now I have 2
PutOutFire(float dousePower) methods in different scripts that are
inpdependant. A bug in one would have to be fixed in both places.
A solution could be to make a base class CanBeStartedOnFire that they both
inherit from. An interface may work as well, but might also result in just
copy pasted code.

Victim:
Reordered calling ToolFinished() and base.Drop() in the Drop method.
ToolFinished would cause the player to call victim.Drop() and make an
infinite loop.