Player Catches Fire
The player will catch fire if they run into a fire trigger capsule.

Fixed some infinite loop problems. Starting on fire would call drop on the
tool. The tool dropping would call ToolFinished on the player. The axe
catch code in ToolFinished would then call Tool.Drop again, making an
infinite loop.
Now, When trying to drop, if canDrop() is true, will go to DropTool().
This is the case for the victim and fire extinguisher. In DropTool(),
holdinTool is set to false immediately.
For the axe case, if canDrop() was false, it calls ToolFinished() when it
is done swinging. In ToolFinished(), it checks if it is on fire and if the
player is holding a tool. If these checks pass, it will call Pickup again
to try to drop the tool. The holdingTool check causes ToolFinished() to
fail for the extinguisher and victim  if they have already run DropTool()
to prevent infinite loops.

Victim Prefab
In the last commit where the Tool layer was set up, I didn't hit apply on
the victim prefab, just the extinguisher and axe.