House Interior Rooms
The house building algorithm now builds interior rooms, each with a door
leading to a central hallway.
It makes x rooms on each side of the hallway. Rooms are all the same size
and mirrored on opposite side of the hallway.

Game can be lost if too many of the blocks are destroyed by fire. Loss is
being handled by House, should be handled by gameplay manager instead.