Hose Movement
The player uses Hose.MoveTowards when using the hose. When the player is
using the hose and pressing movement keys, they move normally, but slower.
When using the hose and not pressing movement keys, they get pushed back
like the extinguisher.
Player moves normally when holding the hose, but not using it. This may
cause some problems.

Fixed a bug with extinguisher. Used to be able to drop it while using it.
The extinguisher would keep firing on its own. Now it stops when it is
dropped.