Hose Follow
Having the raycast in FireNozzle didn't work because it could not collide
with itself.

Made a Hose script to handle the hose following the nozzle.

It fires rays to see if it is within line of sight of the nozzle. When it
isn't within line of sight, it adjusts its anchor point. It uses the last
point it saw the nozzle at. It goes along that line the distance of the
new thing it hit. This is an attempt at finding the corner of the objecct
it hit. Hopefully from this new anchor point, it can see the player.

Add a small amount to the distance of the obstruction or it gets stuck.
Essentially multiplying the new line it is trying to make by 0.

Doesn't create a hose or have any sort of hose length limit at this stage.