Axe model
Added a low poly fireman axe model. Already has materials on it in the
correct spot.
Added the model as a child of the axe.
Changed some collider boxes to match the axe visuals. The axe trigger
cheats left a bit so the player breaks blocks in front of them, not in
front of the axe.