Destruction By Fire
Blocks destroyed by fire now have their visuals and colliders turned off.
Blocks have their attached collider and mesh renderer set inactive. They
can also have an optional visuals that will be set inactive when
destroyed. This is the door model for the door. Walls and floor may not
get a door; the unity cube may be good enough.

The rubble model is made visible for the walls and doors when they are
destroyedd by fire.

When the fire has destroyed the block, the fire is just smouldering.
Smouldering fire does not light the player on fire.