Smoulder Particles
When the object is destroyed by fire, the fire particles change. Same
particles as the fire extinguiseher, but a darker colour.

Right when the change happens, all the current flame particles all change
to the smoke particles. This kind of makes a nice poof to know the thing
is destroyed, but also looks kind of bad that the fire all turns to smoke.
Having 2 different particles systems might be better.