Adjusted Smoulder Smoke
The fire particles don't swap to the smoke particles for a second when the
block is destroyed. This gives a chance for the old fire particles to die
off. The new fire particles are slower and don't travel as high. They are
less noticable when they are swapped with smoke particles.