LinkedList of Hose Positions
Added the points where the hose loses sight of the nozzle into a linked
list.

Made a debug loop that shows a line from each point to the next in the
list. The line shows where the hose should be.

Hose cheats a little bit more. Now jumps 0.1 (instead of 0.01) along the
last line it could see when it hits an object.

Still some problems with the height of the raycast. The hose is raycasting
from the ground up towards where the player holds the nozzle. Sometimes it
seems like it can see over top of objects it shouldn't be able to.
Also doesn't ignore objects that it should ignore. Might have to do
raycastAll() to get it to see through other tools.