Hose Max Length and Mesh
The hose tracks the length it has travelled and can tell when the max
length is reached.

Created a mesh for the hose to alter as it moves. Currently only shows the
mesh where the hose had to take a corner. The mesh doesn't continue to the
player from the last point.

The mesh creates 2 new verts and 2 new triangles when it hits an object.
It creates a new triangle and vert array that are 2 larger (technically 6
in the triangles). It loops through the old arrays, filling the new arrays
except the last 2 items. It then calculates the last 2 items and adds them
to the arrays.

This feels like it might be an inefficient manner of doing this. Maybe
making a mesh of ~50 vertices at the start, then just moving the vertices
as the player moves would be better? Only creating new vertices once the
original 50 are used up. Similar to an object pooling system. Would this
mean I would have to update the position of 48 vertices until I can drop
some off at places they will stay? Or could I get away with only moving
the head 2 vertices?