Hose made of HoseSegments
The hose is made of a bunch of hose segements. Each segment tries to
follow the next segment in line. They try to stay a minimum distance away
from one another.

The lead segment should pull back on the nozzle if it gets too far away.

The tail segment should pull the hydrant so it follow behind the hose.

It doesn't really work like that. The lead segment won't follow the
nozzle, just pull it back.