Tweaked Rolling
Distance needed to roll is much smaller.
Need to roll more times to put out fire.
Same overall distance still needed to be covered.

This makes it so you don't get stuck in a hallway when rolling, not being
able to roll far enough to get the roll to count in either direction.