Start Segment Movement
The start segment isn't worried about how far away the fire hydrant anchor
is. The start segment moves as long as it's not too close to the next
segment. When it gets too far away from the hydrant, it pulls it along.

Changed the segments to be trigger colliders. Otherwise, the player can't
move through the segments.

Segments now follow behind the player and kind of look like a hose.