Preset House Layout
Can now make house layouts by giving an input image of the layout. Map
colours of the input image to prefabs to instantiate. 1 pixel of the input
image maps to 1 grid square.

Check a bool to swap between modes of house building.

Can also pass in a image to set up furniture, but it hasn't been
implemented yet.