Carpet
Added a carpet texture.
Changed floor tiles to not dissappear when destroyed. They just add the
scorch mark on top of the currrent floor. This makes it so it doesn't just
leave a floating scorch mark on top of the ground plane.

Tweaked the colours of the burnt wood (door and wall rubble) a bit.