Rotation in Alpha
Implemented the rotational information in the alpha channel system.

Ignore any pixel in the input image with a == 0 (this is most of them for
furniture).

Check the 3 colour channels to see if colours match.
Get rotation info from alpha.
200 maps to 0 rotation
209 to 90 degrees
218 to 180 degrees
227 to 270 degrees
Formula is the last 2 digits * 10
2 as first digit to keep the image visible in paint.net. 009 would be
nearly see through.

Gave a parent to 2 wide objects. They are rotated based on their bottom
left space. Instead of being spawned half way between 2 tiles.