Furniture Influences Fire
Furniture can now be set up to reduce the time it takes the tile it is
standing on to catch fire and grow to a large fire.

Furniture can set a variable to reduce the time it takes a block to start
on fire.
Furniture can also set its height and width so it can know what other
tiles to adjust. This takes into account the rotation so it gets the
correct tiles.

BuildingBlock:
Raised the distaantBurnMultiplier. Now it heats up blocks 2 tiles away
more.
Added some randomness to the fire. When burning neighbours, multiply the
burn amount by random number. The random number ranges from 80% to 150%
burn strength. This number is then squared to give a more normal
distribution. This gives a final range from 64% to 225%.
Also added some randomness to the fire growth, Some fires can start at
most halfway towards a large fire. This was also squared to give a more
noraml distribution. The final distribution is to have currentFuelSeconds
start from 0% to 49% of growthTime.
This second bit of randomness proved to spread fire a bit too much so it's
currently commented out. Could be uncommented later for faster fire spread
for higher difficulty.
These 2 pieces of randomess didn't do as much as I'd like to make the fire
appear more random. The way the fire burns all its neighbours tends to
smooth out the fire naturally.

Extinguisher:
Raises the height of the raycasts to be at the height of the particles.
Had some problems where the colliders on some furniture blocked rays when
they shouldn't have (namely the bed).

Increased the time to start on fire and growthTime of the tile flooring.
Carpet and hardwood remain the same.