End Game
When the game ends, it teleport all the players to the front of the house
for a picture in front of the house.

Only teleports and locks their input.

Also raised the height of the unarmed use ability. Where the player can
knock down doors and put out fires without any tools.