End Game Photo
Game takes a screenshot at the end of the game when the camera zooms in.

Camera also zooms out to show the house, then zooms into the players.

Save the screenshot in a new folder, Screenshots. This folder isn't
tracked by git.

Reset player moveInput so players don't keep moving when they get
teleported at the end of the game.