PPE as Tools
PPE = personal Protective Equipment. Fire hats and jackets and stuff for
safety.
Can pick up PPE and move them around and drop them like tools. When you
use them, the player puts them on. They then give the player fire
resistance. They can stand in fire longer before they are set on fire.

When in fire, the player's light time goes down. If it reaches 0, they
light on fire. this light time is increased when ppe is donned. The
lightTime doesn't recover when not in fire (may be changed in the future).
When the player lights on fire and are then put out, this lightTime is
reset.