Victims Talk
Victims say things sometimes. The text is a 3d text that appears in the
world and fadeds out over times.

The 3Dtext has a copy of it behind it to act as a sort of outline. It
doesn't match up very well when the sentences get longer.
The 3Dtext is set to look at the camera. Maye it's possible to get the in
world feel of the 3Dtext in a 2D UI text.