Player Pickup Height Tweaks
Made the player fire the ray to pick up stuff from a bit above center.
This should be a sweet spot that lets the player pick up low objects like
the hat and get higher objects like victims on counters or the bed.

Reduced the height of the box collider on the smaller bed to match the
visuals. I had already done this to the larger bed.

PPE hat now floats higher above the ground.
Victim prefab updated to not float above the ground (updated just one
version of the object last commit, not the prefab).