No Duplicate Help Messages
Added a static field in Victim that tracks the last help message used.
If, when a victim tries to choose a random message, that message was the
last message played, they will choose another one.

Had to remove and re-add the TextMesh component of the 3DText. Having a
font size that is not 0 throws warnings with this non-dynamic font. Once
the font size has been changed, it can no longer be changed back to 0,
hence the need to remove and re-add the component.