Pickup Button Prompts
Made an A and B button sprite.
Created a buttonPrompt script. It controls the visibily of the button
prompt. This includes always being to the right and front of the item, no
matter what orientation you drop it in as well as always having the same
rotation as the camera.

Items you can pick up (tools) have a small A button sprite floating near
them. This is to tell the player they can press a on these items to pick
them up.
The sprite is only active when a player is near it. When all players leave
the area, the button prompt dissappears.
The button prompt is attached to the rigidbody components of the tools.
The rigidbody is already getting hidden when the tool is picked up so this
was just convenient.

B button sprite never shows up. This would be the "Use" prompt. A button
is the "Pickup" prompt.

Started to add some checks to make Victims not spawn on top of furniture.
Multi-block furniture is a bit of an issue. The spawning is a fairly low
priority issue. The player can still pick up the victims in these
positions, some just look weird (which could be a plus anyway).

When creating the new script, Monodevelop was not recognizing it so I had
no autocomplete. Couldn't get it to work. Newer versions of Unity may no
longer support Monodevelop? Swapped to Visual Studio 2017.