Floors Under Doors
Add floor objects as furniture under doors.
When the doors are battered down, they will show the floor and not the
gorund plane.
In the furniture layout image, added spots for the floor tiles under
doors.
When the furniture is placed, these floors are too.
These floors are a different prefabs with no script attached. They are
jsut the floor cube.
This adds a second almost identical prefabs which isn't ideal, but should
be fine in this case.