Destroy Furniture With Axe
Axe can destroy furniture.
Made a Furniture tag. When the axe hits an object with this tage, it
destroys it.
Added this tag to all the furniture prefabs. Made extra cubes children of
the cube with the collider on it. So handles and stuff aren't floating in
the air after.

Curtain doesn't get destroyed by the axe because it has no collider.