Player On Fire Messages
Player says things when they are on fire.
They say things when they start on fire and every few seconds while they
are still on fire to remind the player.

Stole most of the code from the victim. Now if anything needs to be
changed, it needs to be cahnged in both places.

The .vs folder got committed earlier than I noticed it. Now it's just
part of the repo because I'm not 100% sure how to get rid of it and it's
not a big enough deal to deal with it.