Button Prompt on Fire
When the player is set on fire, they get a button prompt to tell them to
press B to stop, drop, and roll.

Button Prompt:
If the button prompt has no sphere collider, turn on by default. This b
button prompt is attached to the fire child of the player. When the fire
is enabled, this is visible. The fire is invisible when the player is not
on fire so the prompt is too. It has no need for a collider.

When players lay down to stop, drop, and roll, they face the other way.
Their head is pointed to the bottom of the screen instead of the top. This
is to hide the weird bottom of their bodies (they have no legs).

removed the debug ability to force a player on fire with middle mouse
button.