Camera Zoom
Camera changes its field of view to try to keep all players on the screen
at all times.

CameraController:
Some things appear to be changed, but it seems to just be because of
autoformatting.

Added a method to Calculate the corners of the screen. This is called at
the start and whenever the FOV is changed. The corners are in relation to
the center of the screen at the ground plane level.

Added a method to adjust the FOV. The gameplayManager computes the min and
max x and z positions for the players and tells the camera. The camerera
sees if any of these points are outside the view. If they are, it
increases FOV to get them back in. There are two thrersholds. If a player
is outside both, the FOV grows, if they are in between, the FOV stays. If
they aren't outside either, the FOV shrinks. This makes is more stable on
the edges (a player could just barely leave and it would grow and shrink
constantly, making the screen jitter).
When a player is outside the bottom of the screen, they don't change the
FOV. They offset the center of the camera instead. Increasing the FOV when
at the bottom of the screen does almost nothing to bring that character
into view. All it does is make a huge area off the top of the screen
visible.

GameManager:
Computer the min and max x and z of all players. Give this info to the
camera to fit them on screen.