Boundaries
The house builds boundaries around it so the players stay near the house.
It creates the boundary object at the given distance from the house, then
stretches them so they make a square around the house. This stretching may
have to go if I end up making a model for the boundaries like a picket
fence. Could just be invisible too.

Player:
Made sure the BuildingBlock is not null. Players also have a fire object
on them. This throws errors.

Camera:
For checking out of view in the top and bottom, it now takes into account
the current bottomOffset. This makes the out of the top work better. It
increases the FOV sooner. Before, half the player's body would be off the
top of the screen before it increased the FOV. This is because a player
will exit the bottom of the screen before a player will exit the top. The
bottom is closer because of the angles.
Also changed the rotation of the camera from 60 to 70. This makes it more
top down and easier to see the top of the screen.