Ground Planes
Made a plane for the hosue foundation. It looks like concrete. You can see
this when the walls are destroyed. This goes directly under the house.
The other plane looks like grass. This covers teh rest of the area and a
bit off the screen.
The foundation is moved to the center of the house and scaled to match it.

Made a couple grass textures. The checkerboard one looks better.