Furniture Burns Up
Fire can destroy furniture.
When a tile is destroyed, it raycasts up. If it hits something tagged with
furniture, it destroys it.
Breaking mulit-block furniture isn't as big of a deal as I thought it
would be. When a tile is destroyed, it typically has fire all around it
too so it doesn't look too weird. It would be better if the models were
modular so they could lose chunks, but oh well.
Now the only problem is that the furniture doesn't leave behind rubble.