Hose Smoothing
Added some techniques to smooth out the hose so not all of the segments
get stuck at max range and then can't move anymore even though it's just
all kinked up.
The technique that works best tries to move the segments towards a
position that is the average of the previous and next positions to try to
smooth out corners.
It works, but it also makes it so the hose goes right through walls very
easily and very often.
Making the segments solid makes them hard to walk over and around.