Solid Hose Segments
Made the hose segments solid. Gave them a rigidbody.
Added in the position averaging again.

This version seems to work the best out of any of them.
One problem is the last segment pushes the nozzle away when it gets close.
also, when dropping the nozzle, sometimes it fires the player away.
Probably for the same reason.

Made the segments spawn lower to the ground so they can be walked over by
players.
Their rigidbodies give them gravity so they fall to the ground anyway.