Segment Queued Smoothing
Segments will only try to take the average position approach if the
segments ahead of it have too.
This means the hose will smooth out closer to the player first. The end of
the hose will stay looped up until the player gets farther away and pulls
it with them.

The end segment (attached to the nozzle) stays farther away from it. So it
doesn't push the nozzle away from it.