Screenshot on Menu
The screenshot taken gets saved to the menu object.
It then gets displayed as a sprite on the menu.

Players spawn further to the right in game. It was causing some issues
spawning inside the triggers of the tools.

There is a button to go back to the menu. Currently, it is always active.
Oooh.... buttons will have to work with controllers....

Menu:
Only move players when menu anim is true. This is set false when a new
scene is loaded.

The buttons on the menu scene start the Building scene. They also set the
house number to build when it loads. The buttons that did create the
houses are gone.