Menu Button Controller Input
The menu buttons work with controller (kinda)
There is a button that does nothing that is the default.

First SElected seems to only work when the scene loads. can't set this
afterwards.

House1 could be the default, but then p1 wouldn't be able to press a to
get their character on screen.