Menu Button
Made a menu button as an inbetween for the menu and buttons

When the scene loads, the buttons want to target a menu method, but it has
been destroyed.
It's easier for the copy menu to set MenuButton's menu to the original

Now you can play multiple games

Have 2 spots for screenshots. One for each level.
They aren't in the best positions, but it's ok.