Axe and Extinguisher Sounds
Added sound effects to the axe and extinguisher

The tool script can add AudioSources to tools.
This way I can control multiple sounds on the same object

The axe has a sound when swinging (was an attempt to make it sound like
the hammers in smash bros brawl).
It also has a sound when hitting something. It ignores when it hits the
floor or else it would make the sound way too much.

The extinguisher has sounds when spraying
It has an intro sound, a sound for when at max spray and a sound to fade
out when it stops spraying.
The intro sound loses volume when reaching max spray. The main spray sound
gains volume as spray strength increases. The main spray loops while the
spray button is pressed. The sound is some noise duplicated and reversed
so the loop is pretty seemless. (the dupe and reverse was done in Audacity
so it's part of the sound, not an effect in game).
When the extinguisher stops, the end sound plays. The main sound and end
sound both fade out over a duration.