Menu Improvements
Fire sound fades out when all fire is out instead of just stopping.
Made the fire sound loop and not play on awake

The position the camera checks to see if it can zoom in is smaller. maybe
that will help?

Moved the title and photos up on the menu screen.
Made the buttons bigger and gave them a red select colour.