Players Join at Menu
The number of players is set in the menu.
Players only spawn if their input was detected in the menu.

Difficulty scales with number of players
More fires spawn every 90 seconds for each player extra player.
More victims are spawned based on the number of players.

The camera has a timed failsafe for the picture mode. If it never gets to
the right position, the if will be true after 10 seconds after the game
ends.