Menu Sound
The player start on fire when they are selected in the menu.
The sound is them starting on fire.

There is still some camera jitter.
Tried to smooth it out by having growFOV and shrinkFOV never both true.
There are still some spots where there is camrera jitter, but it might be
a bit better.
I think the real problem is growing one frame then shrinking the next or
vice versa.